Made various small changes and split texting_interface.jsx into two separate files
